src/main/java/seedu/address/model/PatientRecords.java
package seedu.address.model;

import static java.util.Objects.requireNonNull;

import java.util.List;
import java.util.Optional;

import javafx.collections.ObservableList;
import seedu.address.model.patient.Name;
import seedu.address.model.patient.Patient;
import seedu.address.model.patient.UniquePatientList;

/**
 * Wraps all data at the app level
 * Duplicate patients are not allowed (by .isSamePatient comparison)
 */
public class PatientRecords implements ReadOnlyList<Patient> {

    private final UniquePatientList patients;

    /*
     * The 'unusual' code block below is a non-static initialization block, sometimes used to avoid duplication
     * between constructors. See https://docs.oracle.com/javase/tutorial/java/javaOO/initial.html
     *
     * Note that non-static init blocks are not recommended to use. There are other ways to avoid duplication
     *   among constructors.
     */
    {
        patients = new UniquePatientList();
    }

    public PatientRecords() {}

    /**
     * Creates an CovigentApp using the Patients in the {@code toBeCopied}
     */
    public PatientRecords(ReadOnlyList<Patient> toBeCopied) {
        this();
        resetData(toBeCopied);
    }

    //// list overwrite operations

    /**
     * Replaces the contents of the patient list with {@code patients}.
     * {@code patients} must not contain duplicate patients.
     */
    public void setPatients(List<Patient> patients) {
        this.patients.setPatients(patients);
    }

    /**
     * Resets the existing data of this {@code PatientRecords} with {@code newData}.
     */
    public void resetData(ReadOnlyList<Patient> newData) {
        requireNonNull(newData);

        setPatients(newData.getReadOnlyList());
    }

    //// patient-level operations

    /**
     * Returns true if a patient with the same identity as {@code patient} exists in the patient records.
     */
    public boolean hasPatient(Patient patient) {
        requireNonNull(patient);
        return patients.contains(patient);
    }

    //@@author chiamyunqing
    /**
     * Returns the patient with the input name.
     * @param name of patient to find.
     * @return the patient with the name if it exists.
     */
    public Optional<Patient> getPatientWithName(Name name) {
        requireNonNull(name);
        return patients.getPatientWithName(name);
    }

    //@@author

    /**
     * Adds a patient to the address book.
     * The patient must not already exist in the address book.
     */
    public void addPatient(Patient p) {
        patients.add(p);
    }

    /**
     * Replaces the given patient {@code target} in the list with {@code editedPatient}.
     * {@code target} must exist in the address book.
     * The patient identity of {@code editedPatient} must not be the same as another existing patient
     * in the address book.
     */
    public void setPatient(Patient target, Patient editedPatient) {
        requireNonNull(editedPatient);

        patients.setPatient(target, editedPatient);
    }

    /**
     * Removes {@code key} from this {@code PatientRecords}.
     * {@code key} must exist in the address book.
     */
    public void removePatient(Patient key) {
        patients.remove(key);
    }

    //// util methods

    @Override
    public String toString() {
        return patients.asUnmodifiableObservableList().size() + " patients";
        // TODO: refine later
    }

    @Override
    public ObservableList<Patient> getReadOnlyList() {
        return patients.asUnmodifiableObservableList();
    }

    @Override
    public boolean equals(Object other) {
        return other == this // short circuit if same object
                || (other instanceof PatientRecords // instanceof handles nulls
                && patients.equals(((PatientRecords) other).patients));
    }

    @Override
    public int hashCode() {
        return patients.hashCode();
    }
}
